Update existing material to work with VisiData 2.x
This commit brings the terminal commands and outputs in line with
VisiData 2.x, and updates some of the written material accordingly. It
does not, however, add material that is new to 2.x. Some outdated
information (such as lack of undo in 1.5.2) will need to be updating in
future commits, as well.